fix:change get_on_sell function to consumer service
