+++ backend/models/consumer.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+from backend.models.seller import UploadItemRequest
+class ProductInfo(UploadItemRequest):
+    """
+    the class extends UploadItemRequest:
+    Model representing presented agricultural.
+    """
+    id: int
+    uploadDate: str
